add new DelphesLCIOConverter class
