+++ examples/cpp/delphes2lcio/CMakeLists.txt
@@ -14,7 +14,7 @@ find_package(ROOT)
 find_package(Delphes)
 
 # build the executable and put it into bin/
-add_executable(DelphesSTDHEP2LCIO DelphesSTDHEP2LCIO.cpp LCIOTreeWriter.cc)
+add_executable(DelphesSTDHEP2LCIO DelphesSTDHEP2LCIO.cpp DelphesLCIOConverter.cc)
 target_include_directories(DelphesSTDHEP2LCIO PRIVATE ${DELPHES_INCLUDE_DIRS} ${LCIO_INCLUDE_DIRS})
 target_link_libraries(DelphesSTDHEP2LCIO ${LCIO_LIBRARIES} ${DELPHES_LIBRARY} ROOT::Core ROOT::RIO ROOT::Tree ROOT::Physics  ROOT::GenVector ROOT::MathCore)
 install(TARGETS DelphesSTDHEP2LCIO DESTINATION bin)
